fix issue that Explicit type conversion not allowed from "uint256" to "address".

+++ contracts/lib/RLPDecode.sol
@@ -117,7 +117,7 @@ library RLPDecode {
         // 1 byte for the length prefix
         require(item.len == 21);
 
-        return address(toUint(item));
+        return address(uint160(toUint(item)));
     }
 
     function toUint(RLPItem memory item) internal pure returns (uint) {
